+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.0.2 (Unreleased)
 
+- feat: early exit velocity solving for island when converged
 - feat: rename `kcc.setRotation()` to `kcc.setQuaternion()`
 
 ## v0.0.1